Remove findBy Include offset to findAll Include offset to findAll and find unit tests Update readme
